# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "admin接口文档",
        "title": "admin API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/admin/api/menu/api/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取菜单下的接口",
                "tags": [
                    "menu"
                ],
                "summary": "获取菜单下的接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Menu"
                            }
                        }
                    }
                }
            }
        },
        "/admin/api/menu/authorize": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取菜单权限",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "获取菜单权限",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Menu"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "children": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/models.Menu"
                                                }
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/admin/api/menu/authorize/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新菜单权限",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "更新菜单权限",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAuthorizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/api/menu/bind-api": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "绑定菜单下的接口",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "绑定菜单下的接口",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MenuBindAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/api/menu/tree": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取菜单树",
                "tags": [
                    "menu"
                ],
                "summary": "获取菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "allOf": [
                                    {
                                        "$ref": "#/definitions/models.Menu"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "children": {
                                                "type": "array",
                                                "items": {
                                                    "$ref": "#/definitions/models.Menu"
                                                }
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    }
                }
            }
        },
        "/admin/api/menus": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "菜单列表数据",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "菜单列表数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "parentID",
                        "name": "parentID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Page"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Menu"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建菜单",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Menu"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Menu"
                        }
                    }
                }
            }
        },
        "/admin/api/menus/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取菜单",
                "tags": [
                    "menu"
                ],
                "summary": "获取菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "preloads",
                        "name": "preloads",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Menu"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新菜单",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Menu"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除菜单",
                "tags": [
                    "menu"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.MenuBindAPIRequest": {
            "type": "object",
            "required": [
                "menuID",
                "paths"
            ],
            "properties": {
                "menuID": {
                    "type": "string"
                },
                "paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.UpdateAuthorizeRequest": {
            "type": "object",
            "required": [
                "keys",
                "roleID"
            ],
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roleID": {
                    "type": "string"
                }
            }
        },
...
        "models.Menu": {
            "type": "object",
            "properties": {
                "access": {
                    "description": "Access 权限配置，需要与 plugin-access 插件配合使用",
                    "type": "string"
                },
                "component": {
                    "description": "Component 组件",
                    "type": "string"
                },
                "createdAt": {
                    "description": "CreatedAt create time",
                    "type": "string"
                },
                "fixSiderbar": {
                    "description": "FixedSideBar 固定菜单",
                    "type": "boolean"
                },
                "fixedHeader": {
                    "description": "FixedHeader 固定顶栏",
                    "type": "boolean"
                },
                "flatMenu": {
                    "description": "FlatMenu 子项往上提，仍旧展示",
                    "type": "boolean"
                },
                "footerRender": {
                    "description": "FooterRender 不展示页脚",
                    "type": "boolean"
                },
                "headerRender": {
                    "description": "HeaderRender 不展示顶栏",
                    "type": "boolean"
                },
                "headerTheme": {
                    "description": "HeaderTheme 顶部导航的主题，mix 模式生效",
                    "type": "string"
                },
                "hideChildrenInMenu": {
                    "description": "HideChildrenInMenu 隐藏子菜单",
                    "type": "boolean"
                },
                "hideInBreadcrumb": {
                    "description": "HideInBreadcrumb 在面包屑中隐藏",
                    "type": "boolean"
                },
                "hideInMenu": {
                    "description": "HideInMenu 隐藏自己和子菜单",
                    "type": "boolean"
                },
                "icon": {
                    "description": "Icon 图标",
                    "type": "string"
                },
                "id": {
                    "description": "ID primary key",
                    "type": "string"
                },
                "layout": {
                    "description": "Layout 导航菜单的位置, side 为正常模式，top菜单显示在顶部，mix 两种兼有",
                    "type": "string"
                },
                "menuHeaderRender": {
                    "description": "MenuHeaderRender 不展示菜单头部",
                    "type": "boolean"
                },
                "menuRender": {
                    "description": "MenuRender 不展示菜单",
                    "type": "boolean"
                },
                "method": {
                    "description": "Method 请求方法",
                    "type": "string"
                },
                "name": {
                    "description": "Name 菜单名称",
                    "type": "string"
                },
                "navTheme": {
                    "description": "NavTheme 导航菜单的主题",
                    "type": "string"
                },
                "parentID": {
                    "description": "ParentID 父级id",
                    "type": "string"
                },
                "path": {
                    "description": "// Title 菜单标题\nTitle string `json:\"title\" gorm:\"column:title;comment:菜单标题;type:varchar(255);not null\"`\nPath 路由",
                    "type": "string"
                },
                "permission": {
                    "description": "Permission 菜单权限",
                    "type": "string"
                },
                "sort": {
                    "description": "Sort 排序",
                    "type": "integer"
                },
                "status": {
                    "description": "Status 状态",
                    "type": "string"
                },
                "target": {
                    "description": "Target 新页面打开",
                    "type": "string"
                },
                "type": {
                    "description": "Type 菜单类型",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt update time",
                    "type": "string"
                }
            }
        },
...
        "response.Page": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}